refactor: SortBar 로직 분리
